requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.4
asgiref==3.4.1
asn1crypto==1.5.1
autopep8==1.5.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
Brotli==1.0.9
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
dateparser==1.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-admin-autocomplete-filter==0.7.1
django-admin-filter==0.7
django-admin-numeric-filter==0.1.6
django-admin-tools==0.9.2
django-advanced-filters==1.4.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-easy-audit==1.3.3
django-extra-views==0.14.0
django-fieldsets-with-inlines==0.6
django-filter==22.1
django-formtools==1.0
django-import-export==2.8.0
django-jazzmin==2.5.0
django-rest-framework==0.1.0
django-reversion==5.0.1
django-seed==0.3.1
django-smart-selects==1.5.9
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-summernote==0.8.20.0
django-weasyprint==2.2.0
djangoql==0.17.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
Faker==15.1.1
fonttools==4.37.1
future==0.15.2
google-api-python-client==1.4.1
greenlet==1.1.3
gunicorn==20.1.0
html5lib==1.1
httplib2==0.9.2
idna==3.3
kombu==5.2.4
lxml==4.9.1
MarkupPy==1.14
mssql-django==1.1.3
oauth2client==4.1.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.2.0
ply==3.11
prompt-toolkit==2.0.9
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pydyf==0.2.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pyodbc==4.0.32
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
regex==2022.3.2
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.10.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
tinycss2==1.1.1
toml==0.10.2
toposort==1.7
tornado==6.2
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
Unipath==1.1
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.12
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
wfastcgi==3.0.0
whitenoise==5.3.0
wincertstore==0.2
xhtml2pdf==0.2.8
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zopfli==0.2.1
